throttle: enf: Move filter to throttle pkg
